Module amplify::confinement

Confinement puts a constrain on the number of elements within a collection.

Structs

Confined

The confinement for the collection.

Enums

Error

Errors when confinement constraints were not met.

Traits

Collection

Trait implemented by a collection types which need to support collection confinement.

KeyedCollection

Trait implemented by key-value maps which need to support collection confinement.

Type Aliases

LargeDeque

VecDeque with maximum 2^32-1 items of type T.

LargeHashMap

HashMap with maximum 2^32-1 items.

LargeHashSet

HashSet with maximum 2^32-1 items of type T.

LargeOrdMap

BTreeMap with maximum 2^32-1 items.

LargeOrdSet

BTreeSet with maximum 2^32-1 items of type T.

LargeString

String with maximum 2^32-1 characters.

LargeVec

Vec with maximum 2^32-1 items of type T.

MediumDeque

VecDeque with maximum 2^24-1 items of type T.

MediumHashMap

HashMap with maximum 2^24-1 items.

MediumHashSet

HashSet with maximum 2^24-1 items of type T.

MediumOrdMap

BTreeMap with maximum 2^24-1 items.

MediumOrdSet

BTreeSet with maximum 2^24-1 items of type T.

MediumString

String with maximum 2^24-1 characters.

MediumVec

Vec with maximum 2^24-1 items of type T.

NonEmptyDeque

VecDeque which contains at least a single item.

NonEmptyHashMap

HashMap which contains at least a single item.

NonEmptyHashSet

HashSet which contains at least a single item.

NonEmptyOrdMap

BTreeMap which contains at least a single item.

NonEmptyOrdSet

BTreeSet which contains at least a single item.

NonEmptyString

String which contains at least a single character.

NonEmptyVec

Vec which contains at least a single item.

SmallDeque

VecDeque with maximum 2^16-1 items of type T.

SmallHashMap

HashMap with maximum 2^16-1 items.

SmallHashSet

HashSet with maximum 2^16-1 items of type T.

SmallOrdMap

BTreeMap with maximum 2^16-1 items.

SmallOrdSet

BTreeSet with maximum 2^16-1 items of type T.

SmallString

String with maximum 2^16-1 characters.

SmallVec

Vec with maximum 2^16-1 items of type T.

TinyDeque

VecDeque with maximum 255 items of type T.

TinyHashMap

HashMap with maximum 255 items.

TinyHashSet

HashSet with maximum 255 items of type T.

TinyOrdMap

BTreeMap with maximum 255 items.

TinyOrdSet

BTreeSet with maximum 255 items of type T.

TinyString

String with maximum 255 characters.

TinyVec

Vec with maximum 255 items of type T.