Struct amplify::confinement::Confined

pub struct Confined<C: Collection, const MIN_LEN: usize, const MAX_LEN: usize>(/* private fields */);

The confinement for the collection.

Implementations

impl<C: Collection, const MIN_LEN: usize, const MAX_LEN: usize> Confined<C, MIN_LEN, MAX_LEN>

pub fn try_from(col: C) -> Result<Self, Error>

Tries to construct a confinement over a collection. Fails if the number of items in the collection exceeds one of the confinement bounds.

pub fn try_from_iter<I: IntoIterator<Item = C::Item>>( iter: I ) -> Result<Self, Error>

Tries to construct a confinement with a collection of elements taken from an iterator. Fails if the number of items in the collection exceeds one of the confinement bounds.

pub fn push(&mut self, elem: C::Item) -> Result<(), Error>

Attempts to add a single element to the confined collection. Fails if the number of elements in the collection already maximal.

pub fn extend<T: IntoIterator<Item = C::Item>>( &mut self, iter: T ) -> Result<(), Error>

Attempts to add all elements from an iterator to the confined collection. Fails if the number of elements in the collection already maximal.

pub fn unbox(self) -> C

Removes confinement and returns the underlying collection.

impl<C, const MAX_LEN: usize> Confined<C, ZERO, MAX_LEN>
where C: Default + Collection,

pub fn new() -> Self

Constructs a new confinement containing no elements.

pub fn with_capacity(capacity: usize) -> Self

Constructs a new confinement containing no elements, but with a pre-allocated storage for the capacity of elements.

pub fn clear(&mut self)

Removes all elements from the confined collection.

impl<C, const MAX_LEN: usize> Confined<C, ONE, MAX_LEN>
where C: Default + Collection,

pub fn with(elem: C::Item) -> Self

Constructs a confinement with a collection made of a single required element.

impl<C, const MIN_LEN: usize> Confined<C, MIN_LEN, U8>
where C: Default + Collection,

pub fn len_u8(&self) -> u8

Returns number of elements in the confined collection as u8. The confinement guarantees that the collection length can’t exceed u8::MAX.

impl<C, const MIN_LEN: usize> Confined<C, MIN_LEN, U16>
where C: Default + Collection,

pub fn len_u16(&self) -> u16

Returns number of elements in the confined collection as u16. The confinement guarantees that the collection length can’t exceed u16::MAX.

impl<C, const MIN_LEN: usize> Confined<C, MIN_LEN, U24>
where C: Default + Collection,

pub fn len_u24(&self) -> u24

Returns number of elements in the confined collection as u24. The confinement guarantees that the collection length can’t exceed u24::MAX.

impl<C, const MIN_LEN: usize> Confined<C, MIN_LEN, U32>
where C: Default + Collection,

pub fn len_u32(&self) -> u32

Returns number of elements in the confined collection as u32. The confinement guarantees that the collection length can’t exceed u32::MAX.

impl<C: KeyedCollection, const MIN_LEN: usize, const MAX_LEN: usize> Confined<C, MIN_LEN, MAX_LEN>

pub fn insert( &mut self, key: C::Key, value: C::Value ) -> Result<Option<C::Value>, Error>

Inserts a new value into the confined collection under a given key. Fails if the collection already contains maximum number of elements allowed by the confinement.

impl<C, const MAX_LEN: usize> Confined<C, ONE, MAX_LEN>
where C: Default + KeyedCollection,

pub fn with_key_value(key: C::Key, value: C::Value) -> Self

Constructs a confinement with a collection made of a single required key-value pair.

impl<const MIN_LEN: usize, const MAX_LEN: usize> Confined<String, MIN_LEN, MAX_LEN>

pub fn remove(&mut self, index: usize) -> Result<char, Error>

Removes a single character from the confined string, unless the string doesn’t shorten more than the confinement requirement. Errors otherwise.

impl<T, const MIN_LEN: usize, const MAX_LEN: usize> Confined<Vec<T>, MIN_LEN, MAX_LEN>

pub fn remove(&mut self, index: usize) -> Result<T, Error>

Removes an element from the vector at a given index. Errors if the index exceeds the number of elements in the vector, of if the new vector length will be less than the confinement requirement. Returns the removed element otherwise.

impl<T, const MIN_LEN: usize, const MAX_LEN: usize> Confined<VecDeque<T>, MIN_LEN, MAX_LEN>

pub fn push_from(&mut self, elem: T) -> Result<(), Error>

Prepends an element to the deque. Errors if the new collection length will not fit the confinement requirements.

pub fn push_back(&mut self, elem: T) -> Result<(), Error>

Appends an element to the deque. Errors if the new collection length will not fit the confinement requirements.

pub fn remove(&mut self, index: usize) -> Result<T, Error>

Removes an element from the deque at a given index. Errors if the index exceeds the number of elements in the deque, of if the new deque length will be less than the confinement requirement. Returns the removed element otherwise.

impl<T: Hash + Eq, const MIN_LEN: usize, const MAX_LEN: usize> Confined<HashSet<T>, MIN_LEN, MAX_LEN>

pub fn remove(&mut self, elem: &T) -> Result<bool, Error>

Removes an element from the set. Errors if the index exceeds the number of elements in the set, of if the new collection length will be less than the confinement requirement. Returns if the element was present in the set.

pub fn take(&mut self, elem: &T) -> Result<Option<T>, Error>

Removes an element from the set. Errors if the index exceeds the number of elements in the set, of if the new collection length will be less than the confinement requirement. Returns the removed element otherwise.

impl<T: Ord, const MIN_LEN: usize, const MAX_LEN: usize> Confined<BTreeSet<T>, MIN_LEN, MAX_LEN>

pub fn remove(&mut self, elem: &T) -> Result<bool, Error>

Removes an element from the set. Errors if the index exceeds the number of elements in the set, of if the new collection length will be less than the confinement requirement. Returns if the element was present in the set.

pub fn take(&mut self, elem: &T) -> Result<Option<T>, Error>

Removes an element from the set. Errors if the index exceeds the number of elements in the set, of if the new collection length will be less than the confinement requirement. Returns the removed element otherwise.

impl<K: Hash + Eq, V, const MIN_LEN: usize, const MAX_LEN: usize> Confined<HashMap<K, V>, MIN_LEN, MAX_LEN>

pub fn remove(&mut self, key: &K) -> Result<Option<V>, Error>

Removes an element from the map. Errors if the index exceeds the number of elements in the map, of if the new collection length will be less than the confinement requirement. Returns the removed value otherwise.

impl<K: Ord + Hash, V, const MIN_LEN: usize, const MAX_LEN: usize> Confined<BTreeMap<K, V>, MIN_LEN, MAX_LEN>

pub fn remove(&mut self, key: &K) -> Result<Option<V>, Error>

Removes an element from the map. Errors if the index exceeds the number of elements in the map, of if the new collection length will be less than the confinement requirement. Returns the removed value otherwise.