Trait amplify::confinement::KeyedCollection

pub trait KeyedCollection: Collection<Item = (Self::Key, Self::Value)> {
    type Key: Eq + Hash;
    type Value;

    // Required methods
    fn insert(
        &mut self,
        key: Self::Key,
        value: Self::Value
    ) -> Option<Self::Value>;
    fn remove(&mut self, key: &Self::Key) -> Option<Self::Value>;
}

Trait implemented by key-value maps which need to support collection confinement.

Required Associated Types

type Key: Eq + Hash

Key type for the collection.

type Value

Value type for the collection.

Required Methods

fn insert(&mut self, key: Self::Key, value: Self::Value) -> Option<Self::Value>

Inserts a new value under a key. Returns previous value if a value under the key was already present in the collection.

fn remove(&mut self, key: &Self::Key) -> Option<Self::Value>

Removes a value stored under a given key, returning the owned value, if it was in the collection.

Object Safety

This trait is not object safe.