Struct amplify::flags::FlagVec

pub struct FlagVec(/* private fields */);

Structure holding a given set of features

Implementations

impl FlagVec

pub fn new() -> FlagVec

Constructs a features vector of zero feature flag set

pub fn with_capacity(upto: FlagNo) -> Self

Constructs a features vector of upto feature flag in unset state

pub fn is_empty(&self) -> bool

Detects whether structure contains any flags set

pub fn count_flags(&self) -> u16

Counts number of flags set

pub fn as_inner(&self) -> &[u8]

Returns byte slice representation of the inner data (slice of bytes, with 8 bit flags per each byte)

pub fn from_inner(slice: Vec<u8>) -> Self

Constructs flag vector from inner representation (slice of bytes, with 8 bit flags per each byte)

pub fn shrunk(&self) -> Self

Returns a shrunk copy of the self

pub fn iter(&self) -> AllSet<'_>

Creates an iterator for the current feature flags which have “set” state

pub fn known_iter(&self, known: FlagVec) -> FilteredIter<'_>

Creates iterator over known set of the features

pub fn unknown_iter(&self, known: FlagVec) -> FilteredIter<'_>

Creates iterator over unknown set of the features, i.e. features that do not match flags set in known parameter

pub fn shrink(&mut self) -> bool

Reduces the size of the internal buffer to the smallest capacity required to keep all currently set feature flags. Returns true if resize operation was required, or false otherwise, when the internal buffer already was of the smallest possible size

pub fn is_set(&self, flag_no: FlagNo) -> bool

Returns whether a feature flag with flag_no is set (true or false)

pub fn set(&mut self, flag_no: FlagNo) -> bool

Sets the feature flag with number flag_no. Returns true if the flag was already set and false otherwise (i.e. previous value of the flag)

pub fn unset(&mut self, flag_no: FlagNo) -> bool

Unsets the feature flag with number flag_no. Returns true if the flag was set before and false otherwise (i.e. previous value of the flag)

pub fn invert(&mut self, flag_no: FlagNo)

Inverts the feature flag with number flag_no