Module amplify::num

Custom-sized numeric types

Implementation of a various integer types with custom bit dimension. These includes:

large signed and unsigned integers, named gib int types (256, 512, 1024-bit)

custom sub-8 bit unsigned integers, named *small int types (5-, 6-, 7-bit)

24-bit signed integer.

The functions here are designed to be fast.

Modules

error

hex

Hex encoding and decoding

posit

Structs

i256

Large integer type

i512

Large integer type

i1024

Large integer type

u1

1-bit unsigned integer in the range 0..1. It can be used instead of bool when 1-bit numeric (and not boolean) arithmetic is required

u2

2-bit unsigned integer in the range 0..4

u3

3-bit unsigned integer in the range 0..8

u4

4-bit unsigned integer in the range 0..16

u5

5-bit unsigned integer in the range 0..32

u6

6-bit unsigned integer in the range 0..64

u7

7-bit unsigned integer in the range 0..128

u24

24-bit unsigned integer in the range 0..16_777_216

u256

Large integer type

u512

Large integer type

u1024

Large integer type