Struct amplify::Slice32

pub struct Slice32(/* private fields */);

Wrapper type for all slice-based 256-bit types implementing many important traits, so types based on it can simply derive their implementations.

Type keeps data in little-endian byte order and displays them in the same order (like bitcoin SHA256 single hash type).

Implementations

impl Slice32

pub fn random() -> Self

Generates 256-bit array from bitcoin::secp256k1::rand::thread_rng random number generator

pub fn from_slice(slice: impl AsRef<[u8]>) -> Option<Slice32>

Constructs 256-bit array from a provided slice. If the slice length is not equal to 32 bytes, returns None

pub fn as_slice(&self) -> &[u8]

Returns byte slice representation.

pub fn as_slice_mut(&mut self) -> &mut [u8]

Returns mutable byte slice representation.

pub fn to_vec(&self) -> Vec<u8>

Returns vector representing internal slice data