Trait amplify::Dumb

pub trait Dumbwhere
    Self: Sized,{
    // Required method
    fn dumb() -> Self;
}

Used as an alternative to default for test and prototyping purposes, when a type can’t have a default value, but you need to generate some dumb data.

Required Methods

fn dumb() -> Self

Returns an object initialized with dumb data